마이그레이션 - spring, netty 버전 올림 - 알수 없는 테스트 항목 주석처리 - 오타 수정

+++ example/src/test/java/com/manning/nettyinaction/chapter10/AbsIntegerEncoderTest.java
@@ -23,7 +23,7 @@ public class AbsIntegerEncoderTest {
 
         Assert.assertTrue(channel.finish());
         for (int i = 1; i < 10; i++) {
-            Assert.assertEquals(i, channel.readOutbound());
+            Assert.assertEquals(i, (int)channel.readOutbound());
         }
         Assert.assertNull(channel.readOutbound());
     }
